FW 3.66: Many changes, see changelog

+++ encoder.h
@@ -37,7 +37,7 @@ void encoder_tim_isr(void);
 void encoder_set_counts(uint32_t counts);
 bool encoder_index_found(void);
 
-uint16_t encoder_spi_get_val(void);
+uint32_t encoder_spi_get_val(void);
 uint32_t encoder_spi_get_error_cnt(void);
 float encoder_spi_get_error_rate(void);
 uint32_t encoder_sincos_get_signal_below_min_error_cnt(void);
